+++ src/apis/file/index.ts
@@ -1,2 +1,2 @@
-export { uploadToPresignedUrl } from "./fileUpLoad.api";
+export { uploadToPresignedUrl } from "./fileUpload.api";
 export { issuePresignedUrl } from "./presignedUrl.api";
